+++ Models/DTOs/ProductDTOs/ProductDTOs.cs
@@ -9,8 +9,8 @@ namespace SMS_MVCDTO.Models.DTOs.ProductDTOs
         public string Name { get; set; }
         public string Description { get; set; }
         public double SellingPrice { get; set; }
-        public int Quantity { get; set; }
-        public int ReorderLevel { get; set; }
+        public double Quantity { get; set; }
+        public double ReorderLevel { get; set; }
         public bool IsAvailable { get; set; }
     }
 }
